Our new dynamic duo: Request and Uri
Request uses the $_SERVER parameters embedded in a Collection to
determine all the major parts of a HTTP request. Based upon Slim's
Request but left out all the clutter we don't need for this project.

It uses Uri which took over Context's previous job of URL-parsing
to map the request endpoint. Still a bit more than we need, but much
less complicated than Context's logic before.